Completamento script statistiche, agiornamento gannt cons e diario

+++ 5_Gioco/BrutalBumperCars/Assets/Scripts/Data/CarDistanceData.cs
@@ -4,5 +4,5 @@ using System;
 public class CarDistanceData
 {
     public string carName;
-    public float distance;
+    public long distance;
 }
